63. Unique Paths II/solution.py
# You are given an m x n integer array grid. There is a robot initially located at the top-left corner (i.e., grid[0][0]). The robot tries to move to the bottom-right corner (i.e., grid[m - 1][n - 1]). The robot can only move either down or right at any point in time.

# An obstacle and space are marked as 1 or 0 respectively in grid. A path that the robot takes cannot include any square that is an obstacle.

# Return the number of possible unique paths that the robot can take to reach the bottom-right corner.

# The testcases are generated so that the answer will be less than or equal to 2 * 109.

 

# Example 1:


# Input: obstacleGrid = [[0,0,0],[0,1,0],[0,0,0]]
# Output: 2
# Explanation: There is one obstacle in the middle of the 3x3 grid above.
# There are two ways to reach the bottom-right corner:
# 1. Right -> Right -> Down -> Down
# 2. Down -> Down -> Right -> Right
# Example 2:


# Input: obstacleGrid = [[0,1],[0,0]]
# Output: 1
 

# Constraints:

# m == obstacleGrid.length
# n == obstacleGrid[i].length
# 1 <= m, n <= 100
# obstacleGrid[i][j] is 0 or 1.

from typing import List

class Solution:
    def uniquePathsWithObstacles(self, obstacleGrid: List[List[int]]) -> int:
        m = len(obstacleGrid)
        n = len(obstacleGrid[0])

        # if [0][0] or [m-1][n-1] has obstacle
        # no way
        if obstacleGrid[0][0] == 1 or obstacleGrid[m-1][n-1] == 1:
            return 0

        dp = [[0] * n for _ in range(m)]

        # initial point is 1 variant
        dp[0][0] = 1

        # first column
        obstacle_i = -1
        for i in range(1, m):
            # can access
            if obstacle_i == -1 and obstacleGrid[i][0] == 0:
                dp[i][0] = 1
            # can't acces. or obstacle now or before
            else:
                dp[i][0] = 0
                obstacle_i = i
        
        # first column
        obstacle_j = -1
        for j in range(1, n):
            # can access
            if obstacle_j == -1 and obstacleGrid[0][j] == 0:
                dp[0][j] = 1
            # can't acces. or obstacle now or before
            else:
                dp[0][j] = 0
                obstacle_j = j
        
        # rest of dp
        for i in range(1, m):
            for j in range(1, n):
                if obstacleGrid[i][j] == 1:
                    dp[i][j] = 0
                else:
                    dp[i][j] = dp[i-1][j] + dp[i][j-1]

        return dp[m-1][n-1]
